fix: cast input buffer size to uint32_t for compatibility

+++ src/network/HTTPBase.cpp
@@ -475,7 +475,7 @@ namespace ITKExtension
                  []()
                  { return true; }},
                 input_buffer.data(),             // input buffer
-                input_buffer.size(),             // input buffer size
+                (uint32_t)input_buffer.size(),   // input buffer size
                 HTTP_MAX_HEADER_RAW_SIZE,        // max_header_size_bytes
                 HTTP_MAX_HEADER_COUNT,           // max header count
                 read_body_until_connection_close // read body until connection close
